requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
Django==2.1.2
django-celery-beat==1.3.0
django-celery-results==1.0.4
django-cleanup==2.1.0
django-timezone-field==3.0
djangorestframework==3.9.0
kombu==4.2.1
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.7
redis==2.10.6
six==1.11.0
vine==1.1.4
